# sdks/openapi.yaml
openapi: "3.0.3"
info:
  title: LLMSafeSpace API
  description: |
    LLMSafeSpace is a Kubernetes-first platform for running AI agents securely in isolated sandboxes.
    Every sandbox runs `opencode serve` as a persistent HTTP server with a PVC-backed persistent workspace.
  version: "1.0.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /workspaces/{id}:
    get:
      tags: [workspaces]
      summary: Get a workspace
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [workspaces]
      summary: Rename a workspace
      operationId: renameWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "204":
          description: Workspace renamed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [workspaces]
      summary: Delete a workspace
      operationId: deleteWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "204":
          description: Workspace deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    WorkspaceId:
      name: id
      in: path
      required: true
      description: Workspace ID
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login or /auth/register
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "lsp_ prefixed API key in Bearer format (e.g. Bearer lsp_abc123...)"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Workspace:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        userId:
          type: string
        runtime:
          type: string
        storageSize:
          type: string
        phase:
          type: string
          enum: [Pending, Active, Suspending, Suspended, Resuming, Terminating, Terminated, Failed]
        pvcName:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...